Refactor package name from s3_safe_storage to safe_s3_storage and update references
