+++ README.md
@@ -1,10 +1,10 @@
-# s3-safe-storage
+# safe-s3-storage
 
 S3 tools for uploading files to S3 safely (antivirus check, etc) as well as downloading and deleting files.
 
 ## How To Use
 
 ```
-uv add s3-safe-storage
-poetry add s3-safe-storage
+uv add safe-s3-storage
+poetry add safe-s3-storage
 ```
