+++ safe_s3_storage/exceptions.py
