+++ tests/test_exceptions.py
@@ -2,7 +2,7 @@ import typing
 
 import faker
 
-from s3_safe_storage.exceptions import ThreatDetectedError
+from safe_s3_storage.exceptions import ThreatDetectedError
 from tests.conftest import generate_binary_content
